> FINISHED: Search Page (Only needs the search form bg photo) > FINISHED: Home Page > About page fixed
