+++ src/components/aboutpage-components/TeamSection.css
@@ -35,7 +35,7 @@
   padding-top: 10px;
 }
 
-a {
+.teamSectionStyle a {
   font-family: "AvenirLTStd-Book";
   color: #79158f;
   padding-left: 50px;
