+++ src/components/aboutpage-components/TopContent.css
@@ -1,4 +1,4 @@
-.container-fluid{
+.aboutFluidContainer {
   width: 1300px;
   padding-top: 60px;
 }
