--- src/components/homepage-components/NewPractitionerCard.css
@@ -1,48 +1,0 @@
-.practitionerCardStyle {
-  margin: 0.5% 15% 0.5% 15%;
-  font-size: 1em;
-  font-family: "AvenirLTStd-Book";
-  position: relative;
-  display: inline-block;
-}
-
-.practitionerCardStyle .card-title {
-  color: #79158f !important;
-  font-size: 1.3vw;
-  font-family: "AvenirLTPro-Heavy";
-}
-
-.practitionerCardStyle .card-subtitle {
-  color: gray !important;
-  font-size: 0.9vw;
-  margin-bottom: 15px !important;
-}
-
-.practitionerCardStyle .card-text {
-  color: gray !important;
-  font-size: 0.8vw;
-}
-
-.practitionerCardStyle:before {
-  content: "";
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  left: 0;
-  top: 0;
-  background: linear-gradient(transparent 320px, white);
-}
-
-.card-img-overlay {
-  right: 0;
-  padding: 0;
-}
-
-#practitionerCardType {
-  color: white;
-  background-color: #79158f;
-  padding: 5px 15px 5px 15px;
-  font-size: 0.8vw;
-  height: 0;
-  top: 47%;
-}
