--- src/components/homepage-components/NewPractitionerList.css
@@ -1,12 +1,0 @@
-.practitionerContainerStyle {
-  padding-top: 4%;
-  padding-bottom: 4%;
-  background-color: lavender;
-}
-
-.practitionerListStyle h4 {
-  margin-bottom: 4%;
-  font-family: "AvenirLTPro-Heavy";
-  font-size: 1.3vw;
-  font-weight: 500;
-}
