+++ src/components/practitioner-components/NewPractitionerList.css
@@ -1,7 +1,7 @@
 .newPracContainerStyle {
   padding-top: 4%;
   padding-bottom: 4%;
-  background-color: lavender;
+  background-color: #F3F1F3;
 }
 
 .newPracListStyle h4 {
@@ -9,6 +9,7 @@
   font-family: "AvenirLTPro-Heavy";
   font-size: 1.3vw;
   font-weight: 500;
+  color: #5B5B5B;
 }
 
 .newPracListStyle a {
