+++ src/components/practitioner-components/NewPractitionerList.js
@@ -2,7 +2,6 @@ import "./NewPractitionerList.css";
 import React from "react";
 import NewPractitionerCard from "./NewPractitionerCard";
 import { Container, CardDeck } from "react-bootstrap";
-import Link from "react-router-dom";
 
 class NewPractitionerList extends React.Component {
   render() {
@@ -40,7 +39,7 @@ class NewPractitionerList extends React.Component {
             />
           </CardDeck>
           <br />
-          <a to="/new-practitioner-list">Show All (5)</a>
+          <a href="./new-practitioner-list">Show All (5)</a>
         </Container>
       </Container>
     );
