+++ src/components/practitioner-components/PractitionerList.css
@@ -1,7 +1,7 @@
 .pracContainerStyle {
   padding-top: 4%;
   padding-bottom: 4%;
-  background-color: lavender;
+  background-color: #F3F1F3;
 }
 
 .pracListStyle h4 {
@@ -9,6 +9,7 @@
   font-family: "AvenirLTPro-Heavy";
   font-size: 1.3vw;
   font-weight: 500;
+  color: #5B5B5B;
 }
 
 .pracListStyle a {
@@ -16,4 +17,8 @@
   font-weight: 500;
   font-size: 0.9vw;
   color: #79158f;
+}
+
+.pracListDeckStyle {
+  margin: 0 !important;
 }
