+++ src/components/search-components/TherapyList.css
@@ -1,0 +1,26 @@
+.therapyListStyle {
+  margin-top: 3%;
+  margin-bottom: 3%;
+}
+
+.therapyListStyle h4 {
+  font-family: "AvenirLTPro-Heavy";
+  color: #5b5b5b;
+  margin-bottom: 30px;
+}
+
+.therapyListStyle ul {
+  list-style-type: none;
+  padding-left: 0;
+}
+
+.therapyListStyle li {
+  font-family: "AvenirLTStd-Book";
+  font-weight: 500;
+  font-size: 1.2em;
+  color: #79158f;
+}
+
+.therapyListStyle .col-sm-3 {
+  padding-left: 0;
+}
